move the feature modules
