+++ tests/feature/test_atomic_contact.py
@@ -1,11 +1,13 @@
+from uuid import uuid4
+
 from pdb2sql import pdb2sql
 import numpy
 
 from deeprank_gnn.models.structure import Chain, Atom
 from deeprank_gnn.models.contact import AtomicContact, ResidueContact
-from deeprank_gnn.models.graph import Edge
+from deeprank_gnn.models.graph import Edge, Graph
 from deeprank_gnn.tools.pdb import get_structure
-from deeprank_gnn.feature.edge.atomic_contact import add_features
+from deeprank_gnn.feature.atomic_contact import add_features
 from deeprank_gnn.domain.feature import FEATURENAME_EDGEDISTANCE, FEATURENAME_EDGEVANDERWAALS, FEATURENAME_EDGECOULOMB
 
 
@@ -19,6 +21,11 @@ def _get_atom(chain: Chain, residue_number: int, atom_name: str) -> Atom:
 
     raise ValueError(f"Not found: chain {chain.id} residue {residue_number} atom {atom_name}")
 
+
+def _wrap_in_graph(edge: Edge):
+    g = Graph(uuid4().hex)
+    g.add_edge(edge)
+    return g
 
 def test_add_features():
 
@@ -34,7 +41,7 @@ def test_add_features():
     contact = AtomicContact(_get_atom(structure.chains[0], 0, "N"),
                             _get_atom(structure.chains[0], 0, "CA"))
     edge_close = Edge(contact)
-    add_features(pdb_path, [edge_close])
+    add_features(pdb_path, _wrap_in_graph(edge_close))
     assert not numpy.isnan(edge_close.features[FEATURENAME_EDGEVANDERWAALS])
     assert edge_close.features[FEATURENAME_EDGEVANDERWAALS] > 0.0, edge_close.features[FEATURENAME_EDGEVANDERWAALS]
 
@@ -42,7 +49,7 @@ def test_add_features():
     contact = AtomicContact(_get_atom(structure.chains[0], 0, "N"),
                             _get_atom(structure.chains[0], 27, "CB"))
     edge_far = Edge(contact)
-    add_features(pdb_path, [edge_far])
+    add_features(pdb_path, _wrap_in_graph(edge_far))
     assert not numpy.isnan(edge_far.features[FEATURENAME_EDGEVANDERWAALS])
     assert edge_far.features[FEATURENAME_EDGEVANDERWAALS] < 0.0, edge_far.features[FEATURENAME_EDGEVANDERWAALS]
 
@@ -50,7 +57,7 @@ def test_add_features():
     contact = AtomicContact(_get_atom(structure.chains[0], 0, "N"),
                             _get_atom(structure.chains[0], 138, "CG"))
     edge_intermediate = Edge(contact)
-    add_features(pdb_path, [edge_intermediate])
+    add_features(pdb_path, _wrap_in_graph(edge_intermediate))
     assert not numpy.isnan(edge_intermediate.features[FEATURENAME_EDGEVANDERWAALS])
     assert edge_intermediate.features[FEATURENAME_EDGEVANDERWAALS] < edge_far.features[FEATURENAME_EDGEVANDERWAALS], "{} >= {}".format(edge_intermediate.features[FEATURENAME_EDGEVANDERWAALS],
                                                                                                                                        edge_far.features[FEATURENAME_EDGEVANDERWAALS])
@@ -65,7 +72,7 @@ def test_add_features():
     contact = AtomicContact(_get_atom(structure.chains[0], 139, "CZ"),
                             _get_atom(structure.chains[0], 136, "OE2"))
     close_attracting_edge = Edge(contact)
-    add_features(pdb_path, [close_attracting_edge])
+    add_features(pdb_path, _wrap_in_graph(close_attracting_edge))
     assert not numpy.isnan(close_attracting_edge.features[FEATURENAME_EDGECOULOMB])
     assert close_attracting_edge.features[FEATURENAME_EDGECOULOMB] < 0.0, close_attracting_edge.features[FEATURENAME_EDGECOULOMB]
 
@@ -73,7 +80,7 @@ def test_add_features():
     contact = AtomicContact(_get_atom(structure.chains[0], 139, "CZ"),
                             _get_atom(structure.chains[0], 20, "OD2"))
     far_attracting_edge = Edge(contact)
-    add_features(pdb_path, [far_attracting_edge])
+    add_features(pdb_path, _wrap_in_graph(far_attracting_edge))
     assert not numpy.isnan(far_attracting_edge.features[FEATURENAME_EDGECOULOMB])
     assert far_attracting_edge.features[FEATURENAME_EDGECOULOMB] < 0.0, far_attracting_edge.features[FEATURENAME_EDGECOULOMB]
     assert far_attracting_edge.features[FEATURENAME_EDGECOULOMB] > close_attracting_edge.features[FEATURENAME_EDGECOULOMB], "{} <= {}".format(far_attracting_edge.features[FEATURENAME_EDGECOULOMB],
@@ -83,14 +90,14 @@ def test_add_features():
     contact = AtomicContact(_get_atom(structure.chains[0], 109, "OE2"),
                             _get_atom(structure.chains[0], 105, "OE1"))
     opposing_edge = Edge(contact)
-    add_features(pdb_path, [opposing_edge])
+    add_features(pdb_path, _wrap_in_graph(opposing_edge))
     assert not numpy.isnan(opposing_edge.features[FEATURENAME_EDGECOULOMB])
     assert opposing_edge.features[FEATURENAME_EDGECOULOMB] > 0.0, opposing_edge.features[FEATURENAME_EDGECOULOMB]
 
     # check that we can calculate residue contacts
     contact = ResidueContact(structure.chains[0].residues[0], structure.chains[0].residues[1])
     edge = Edge(contact)
-    add_features(pdb_path, [edge])
+    add_features(pdb_path, _wrap_in_graph(edge))
     assert not numpy.isnan(edge.features[FEATURENAME_EDGEDISTANCE]) > 0.0
     assert edge.features[FEATURENAME_EDGEDISTANCE] > 0.0
     assert edge.features[FEATURENAME_EDGEDISTANCE] < 1e5
